squashes evan-flir-kinetic changes
avoids mistakenly having melodic branch commits in the history
