+++ autorally_control/src/path_integral/mppi_nodelet.cu
@@ -34,4 +34,4 @@ namespace autorally_control{
   typedef MPPI<Controller, DynamicsModel, MPPICosts> MPPINodelet;
 }
 
-PLUGINLIB_DECLARE_CLASS(autorally_control, MPPINodelet, autorally_control::MPPINodelet, nodelet::Nodelet)
+PLUGINLIB_EXPORT_CLASS(autorally_control::MPPINodelet, nodelet::Nodelet)
