+++ autorally_core/launch/camera_ptgrey.launch
@@ -55,7 +55,7 @@
     <remap from="camera/image_display" to="$(arg side)_camera/image_display" />
   </node>
   
-  <include file="$(find autorally_core)/launch/camera_auto_balance.launch">
+  <include file="$(find autorally_core)/launch/camera_auto_balance_ptgrey.launch">
     <arg name="serial" value="$(arg serial)" />
     <arg name="side" value="$(arg side)" />
   </include>
