+++ autorally_core/src/CameraAutoBalance/CMakeLists.txt
@@ -1,18 +1,41 @@
-
 find_library(FLYCAP_LIB flycapture)
-
 if(FLYCAP_LIB)
-  add_library(CameraAutoBalance CameraAutoBalance.cpp)
-  add_dependencies(CameraAutoBalance autorally_msgs_gencpp)
-  set_target_properties(CameraAutoBalance PROPERTIES COMPILE_FLAGS "-march=native -O3 -ffast-math")
-  target_link_libraries(CameraAutoBalance ${catkin_LIBRARIES} ${OpenCV_LIBRARIES} ${cv_bridge_LIBRARIES} ${FLYCAP_LIB})
+  add_library(FlycapAdjuster FlycaptureAdjuster.cpp)
+  target_link_libraries(FlycapAdjuster ${FLYCAP_LIB})
 
-install(TARGETS
-  CameraAutoBalance
-  ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-  LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-  RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
-)
+  add_library(CameraAutoBalancePtGrey CameraAutoBalancePtGrey.cpp)
+  add_dependencies(CameraAutoBalancePtGrey autorally_msgs_gencpp FlycapAdjuster)
+  set_target_properties(CameraAutoBalancePtGrey PROPERTIES COMPILE_FLAGS "-march=native -O3 -ffast-math")
+  target_link_libraries(CameraAutoBalancePtGrey FlycapAdjuster ${catkin_LIBRARIES} ${OpenCV_LIBRARIES} ${cv_bridge_LIBRARIES})
+
+  install(TARGETS
+    CameraAutoBalancePtGrey
+    ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
+    LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
+    RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
+  )
 else()
-  message(WARNING "Could not find FlyCapture library. Will not build CameraAutoBalance")
+  message(WARNING "Could not find FlyCapture library. Will not build CameraAutoBalance for Point Grey cameras")
+endif()
+
+
+find_library(Spinnaker_LIB Spinnaker)
+if(Spinnaker_LIB)
+  include_directories(/usr/include/spinnaker /usr/local/include/spinnaker)
+  add_library(SpinnakerAdjuster SpinnakerAdjuster.cpp)
+  target_link_libraries(SpinnakerAdjuster ${Spinnaker_LIB})
+
+  add_library(CameraAutoBalanceFLIR CameraAutoBalanceFLIR.cpp)
+  add_dependencies(CameraAutoBalanceFLIR autorally_msgs_gencpp SpinnakerAdjuster)
+  set_target_properties(CameraAutoBalanceFLIR PROPERTIES COMPILE_FLAGS "-march=native -O3 -ffast-math")
+  target_link_libraries(CameraAutoBalanceFLIR SpinnakerAdjuster ${catkin_LIBRARIES} ${OpenCV_LIBRARIES} ${cv_bridge_LIBRARIES})
+
+  install(TARGETS
+    CameraAutoBalanceFLIR
+    ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
+    LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
+    RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
+  )
+else()
+  message(WARNING "Could not find Spinnaker library. Will not build CameraAutoBalance for FLIR cameras")
 endif()
