+++ autorally_core/src/Diagnostics/CMakeLists.txt
@@ -1,8 +1,8 @@
-add_library(Diagnostics Diagnostics.cpp)
-add_dependencies(Diagnostics autorally_msgs_gencpp)
+add_library(ar_diagnostics Diagnostics.cpp)
+add_dependencies(ar_diagnostics autorally_msgs_gencpp)
 
 install(TARGETS
-  Diagnostics
+  ar_diagnostics
   ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
