+++ autorally_core/src/gps/CMakeLists.txt
@@ -1,12 +1,12 @@
 add_executable(gpsHemisphereInterface GPSHemisphere.cpp)
-target_link_libraries(gpsHemisphereInterface ${catkin_LIBRARIES} SerialSensorInterface Diagnostics)
+target_link_libraries(gpsHemisphereInterface ${catkin_LIBRARIES} SerialSensorInterface ar_diagnostics)
 add_dependencies(gpsHemisphereInterface autorally_msgs_gencpp)
 
 #add_library(gps_LS20031 src/gps_LS20031/gps_LS20031.cpp)
 #add_dependencies(gps_LS20031 autorally_msgs_gencpp)
 
 #add_executable(gpsInterfaceLS src/gpsInterface/gpsInterfaceLS.cpp)
-#target_link_libraries(gpsInterfaceLS ${catkin_LIBRARIES} gps_LS20031 SerialSensorInterface Diagnostics)
+#target_link_libraries(gpsInterfaceLS ${catkin_LIBRARIES} gps_LS20031 SerialSensorInterface ar_diagnostics)
 
 install(TARGETS
   gpsHemisphereInterface
