+++ autorally_core/src/xbee/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_executable(xbeeCoordinator XbeeCoordinator.cpp XbeeInterface.cpp)
-target_link_libraries(xbeeCoordinator ${catkin_LIBRARIES} ${Boost_LIBRARIES} SerialSensorInterface Diagnostics)
+target_link_libraries(xbeeCoordinator ${catkin_LIBRARIES} ${Boost_LIBRARIES} SerialSensorInterface ar_diagnostics)
 add_executable(xbeeNode XbeeNode.cpp XbeeInterface.cpp)
-target_link_libraries(xbeeNode ${catkin_LIBRARIES} ${Boost_LIBRARIES} SerialSensorInterface Diagnostics)
+target_link_libraries(xbeeNode ${catkin_LIBRARIES} ${Boost_LIBRARIES} SerialSensorInterface ar_diagnostics)
 
 install(TARGETS
   xbeeCoordinator
