+++ autorally_core/test/CMakeLists.txt
@@ -1,9 +1,9 @@
 SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -pthread")
 rosbuild_add_gtest(test/diagnosticsTest diagnosticsTest.cpp)
-target_link_libraries(test/diagnosticsTest Diagnostics)
+target_link_libraries(test/diagnosticsTest ar_diagnostics)
 
 rosbuild_add_gtest(test/serialSensorInterfaceTest serialSensorInterfaceTest.cpp)
-target_link_libraries(test/serialSensorInterfaceTest SerialSensorInterface Diagnostics)
+target_link_libraries(test/serialSensorInterfaceTest SerialSensorInterface ar_diagnostics)
 
 rosbuild_add_gtest(test/pololuMicroMaestroTest pololuMicroMaestroTest.cpp)
-target_link_libraries(test/pololuMicroMaestroTest PololuMicroMaestro SerialSensorInterface Diagnostics)
+target_link_libraries(test/pololuMicroMaestroTest PololuMicroMaestro SerialSensorInterface ar_diagnostics)
